+++ lib/src/components/tabs.dart
@@ -77,6 +77,8 @@ class ShadTabs<T> extends StatefulWidget implements PreferredSizeWidget {
     this.value,
     required this.tabs,
     this.controller,
+    this.tabBarAlignment,
+    this.tabsGap,
     this.gap,
     this.scrollable,
     this.dragStartBehavior,
@@ -107,6 +109,16 @@ class ShadTabs<T> extends StatefulWidget implements PreferredSizeWidget {
   /// The controller of the tabs.
   /// {@endtemplate}
   final ShadTabsController<T>? controller;
+
+  /// {@template ShadTabs.tabsGap}
+  /// The horizontal gap between the tabs in the tabBar.
+  /// {@endtemplate}
+  final double? tabsGap;
+
+  /// {@template ShadTabs.tabBarAlignment}
+  /// The alignment of the tabBar.
+  /// {@endtemplate}
+  final Alignment? tabBarAlignment;
 
   /// {@template ShadTabs.gap}
   /// The gap between the tabBar and the content.
@@ -287,7 +299,10 @@ class ShadTabsState<T> extends State<ShadTabs<T>> with RestorationMixin {
     final effectiveExpandContent =
         widget.expandContent ?? tabsTheme.expandContent ?? false;
 
-    Widget tabBar = Row(children: widget.tabs);
+    Widget tabBar = Row(
+      spacing: widget.tabsGap ?? tabsTheme.tabsGap ?? 0,
+      children: widget.tabs,
+    );
 
     if (effectiveTabBarConstraints != null) {
       tabBar = ConstrainedBox(
@@ -311,6 +326,12 @@ class ShadTabsState<T> extends State<ShadTabs<T>> with RestorationMixin {
       );
     } else {
       tabBar = Padding(padding: effectivePadding, child: tabBar);
+    }
+
+    final effectiveTabBarAlignment =
+        widget.tabBarAlignment ?? tabsTheme.tabBarAlignment;
+    if (effectiveTabBarAlignment != null) {
+      tabBar = Align(alignment: effectiveTabBarAlignment, child: tabBar);
     }
 
     return ShadProvider(
@@ -750,12 +771,13 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
     final effectiveBackgroundColor = widget.backgroundColor ??
         tabsTheme.tabBackgroundColor ??
         const Color(0x00000000);
+
     final effectiveSelectedBackgroundColor = widget.selectedBackgroundColor ??
         tabsTheme.tabSelectedBackgroundColor ??
         theme.colorScheme.background;
 
     final effectiveHoverBackgroundColor = widget.hoverBackgroundColor ??
-        tabsTheme.tabHoverForegroundColor ??
+        tabsTheme.tabHoverBackgroundColor ??
         effectiveBackgroundColor;
 
     final effectiveSelectedHoverBackgroundColor =
@@ -800,6 +822,9 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
         widget.pressedBackgroundColor ?? tabsTheme.tabPressedBackgroundColor;
     final effectivePressedForegroundColor =
         widget.pressedForegroundColor ?? tabsTheme.tabPressedForegroundColor;
+
+    final effectiveTextStyle =
+        widget.textStyle ?? tabsTheme.tabTextStyle ?? theme.textTheme.small;
 
     Widget tab = ListenableBuilder(
       listenable: inherited.controller,
@@ -849,7 +874,11 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
         };
 
         final effectiveDecoration = defaultDecoration
-            .merge(tabsTheme.tabDecoration)
+            .merge(
+              selected
+                  ? tabsTheme.tabSelectedDecoration ?? tabsTheme.tabDecoration
+                  : tabsTheme.tabDecoration,
+            )
             .merge(widget.decoration);
 
         return ShadButton.secondary(
@@ -865,9 +894,7 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
               ? effectiveSelectedHoverBackgroundColor
               : effectiveHoverBackgroundColor,
           padding: effectivePadding,
-          decoration: selected
-              ? effectiveDecoration.merge(widget.selectedDecoration)
-              : effectiveDecoration,
+          decoration: effectiveDecoration,
           foregroundColor: selected
               ? effectiveSelectedForegroundColor
               : effectiveForegroundColor,
@@ -907,8 +934,8 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
           onFocusChange: (focused) {
             if (focused) onMayChanged();
           },
-          child: DefaultTextStyle(
-            style: theme.textTheme.small,
+          child: DefaultTextStyle.merge(
+            style: effectiveTextStyle,
             child: widget.child,
           ),
         );